# backend/internal/swagger/swagger.yaml
basePath: /api
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: invalid email or password
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: securePass123
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  dto.RefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: alice@example.com
        maxLength: 255
        type: string
      password:
        example: securePass123
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterResponse:
    properties:
      email:
        example: alice@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/lam-thinh/customer-oil-change-management
  description: A production-ready REST API for managing customer oil change records.
  license:
    name: MIT
  title: Customer Oil Change Management API
  version: "1.0"
paths:
  /v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password; returns an access token and
        sets a secure cookie containing a refresh token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Access token returned and cookie set
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Malformed JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login
      tags:
      - auth
  /v1/logout:
    post:
      consumes:
      - application/json
      description: Revoke a refresh token from cookie. Requires a valid JWT access
        token in the Authorization header.
      produces:
      - application/json
      responses:
        "204":
          description: Logged out and cookie cleared
        "401":
          description: Not authenticated or token not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /v1/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token cookie for a new access token (old
        refresh token is revoked)
      produces:
      - application/json
      responses:
        "200":
          description: New access token returned and cookie set
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "401":
          description: Cookie not found, token revoked, or expired
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /v1/register:
    post:
      consumes:
      - application/json
      description: Create a new account with email and password
      parameters:
      - description: Registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Malformed JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token in the format: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"